requirements.txt
netcdf4 = ""
ipywidgets = ""
click = ""
ipykernel = ""
xarray = ""
numpyro = ""
jax2torch = ""
torch = ""
sbi = ">=0.21.0"
zuko = ""
